Removed Value OverflowException as this was not the purpose of max values for stocks or components.  Components should reach their max value and hold at that value, not throw an exception.
Cloud was entirely rewritten.  The infinity functionality did not exist and the inheritance was entirely flawed.

Component was give add and subtract functionalities so a another component can modify the currentvalues of a component.  This came from a bug discovered in the Flow component.

Control:  Refactored Function evaluation and added a new function type called "recursive".  recursive functions must have an initial value or there is an error.   Added initialization functionality as some values must be initialized based on other components.  This poses a problem common in graph traversal.  A function cannot be initialzed if it relies on components that themselves rely on that component.  To solve this problem I recursively traverse all unitliazed components and their dependencies.  While initializing, a component is flagged.  If a component that is already initializing is asked to be initialized, this repesents a circular reference and the therefore an error indesign.  This will throw an exception that is currently unhandled.

Flow:  modified flow to keep a current value (the value being passed between its source and sink.  also updated flow to utilize new add and subtract methods for the component class.

Stock:  enabled the ability to have a maximum value.  If a max value is reached the stock will simply fill up and the additional values will "disappear"  this poses a design flaw when having to deal with waste or overflow.  Will discuss this issue with Dr. Mobus.

Loader:
  added a control initilzation function and added better error handling for nodes that are optional.
  
SimulationDataStructure:  Only changes to support bug fixes otherwise mentioned.

Config.xml  made adjustments to inititalValues
Config.xsd  made nonrequired fields optional and added required tags to those that are required.  Also changed names from "aStock" to "Stock" etc.